+++ app/src/main/java/com/knu/cloud/data/auth/AuthRemoteDataSource.kt
@@ -11,11 +11,11 @@ class AuthRemoteDataSource @Inject constructor (
     private val authApiService : AuthApiService
 ) {
     suspend fun login(loginRequest: LoginRequest): NetworkResult<AuthResponse<Token>> {
-        Timber.tag("network").d("remoteDataSource login fun 호출")
+        Timber.tag("network").d("AuthRemoteDataSource login fun 호출")
         return authApiService.login(loginRequest)
     }
     suspend fun signUp(signUpRequest: SignUpRequest) : NetworkResult<AuthResponse<Token>> {
-        Timber.tag("network").d("remoteDataSource signUp fun 호출")
+        Timber.tag("network").d("AuthRemoteDataSource signUp fun 호출")
         return authApiService.signUp(signUpRequest)
     }
 }
