+++ app/src/main/java/com/knu/cloud/repository/AuthRepositoryImpl.kt
@@ -6,7 +6,7 @@ import com.knu.cloud.di.*
 import com.knu.cloud.model.NetworkResult
 import com.knu.cloud.model.auth.LoginRequest
 import com.knu.cloud.model.auth.SignUpRequest
-import com.knu.cloud.utils.RetrofitFailureStateException
+import com.knu.cloud.network.RetrofitFailureStateException
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.flow.*
 import javax.inject.Inject
